# swagger.yaml
openapi: 3.0.0
info:
  title: JobsAPi
  contact: {}
  version: "1.0"
servers:
  - url: https://nodejobsapi.herokuapp.com/api/v1
    variables: {}
paths:
...
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
        - Jobs
      summary: GET SINGLE JOB
      operationId: GETSINGLEJOB
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []

    patch:
      tags:
        - Jobs
      summary: UPDATE JOB
      operationId: UPDATEJOB
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UPDATEJOBREQUEST"
                - example:
                    company: netflix and chill
                    position: back-end developer
            example:
              company: netflix and chill
              position: back-end developer
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []

    delete:
      tags:
        - Jobs
      summary: DELETE JOB
      operationId: DELETEJOB
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
...
    UPDATEJOBREQUEST:
      title: UPDATEJOBREQUEST
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: netflix and chill
        position: back-end developer
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...